game: store shield data as array of flags
The flags indicate the state (1 = intact, 0 = destroyed).
The position in the array can be used to determine location x, y.

Whenever GameEntity functions need to be called on the shields, just
create a temporary GameEntity for it on the fly.  This saves a ton of
memory.

Game data memory on AVR: 2533 -> 1669 bytes