Remove game dependency on image filenames
Eventually, there won't be files on the embedded target.  At least not
PNG files.  So eliminate filename-to-index lookup in class Game.  Instead
store images as <type, index>, where type is the game entity type.  This
means only the index of the images of |type| has to be stored.

Example:
  alien-1-3.png -> type=ALIEN, index=2

When the entity's draw() func schedules a blit, it passes <type, index>
instead of the image file index.  The images library will internally look
up the image based on the pair <type, index>.  Thus, all the image
filenames can be removed from Game.

Also, Game no longer needs to store a pointer to the image library.