game: Assign alien indexes based on each alien type
Previously all three types of aliens shared the same index space, and the
index was the position/index within the alien array.

With this patch, each alien type has its own index space from 0 to N-1.

The previous use of alien index has been updated to use the for loop
index.  Storing the alien position as a member variable was a relic left
over from when this game used variable arrays.