game entity: Update usage of alive and active flags
For aliens, "active" means they're firing at you.
For other entities, "active" is analogous to "alive" for aliens.

Since the drawing of entities is governed by the "alive" flag, it should
be set/cleared for non-alien entities that are (in)active.

This patch updates (de)activate() to set the "alive" flag to the same
value as the "active" flag.  This takes care of bonus, explosions, shots,
etc, that had their "alive" flags set even after being deactivated, and
were still being drawn.

For aliens, there is a special initializer where the alive flag can be set
even if the active flag is not.

Note that this patch leaves the game logic unchanged.  It changes the
internal mechanics of the game entities without changing the external
semantics of activate/deactivate.