game: Fix alien speed increase as they get killed
This was broken for the longest time because the fixed point type didn't
have enough precision to represent the speed increase.  Using fixed32,
there is enough precision.

To get this working, |current_alien_speed| had to be turned into a fixed
variable.

And when multiplying it by |delta|, cast |delta| to a signed int.